Handle BK/ZK errors during automatic transaction rollback in clients.
In case of BK/ZK error the client is not able to commit a transaction
it blindly issues a rollback command.
This command may go into a new ledger in case BookKeeperCommitLog
does not closes itself in time.

The fix is to automatically in-memory rollback transactions with a failed
write on the commit command.
This prevents any further command in the scope of the transaction
and expectially rollbacks to be validated.

The problem may occour even on FileCommitLog (standalone mode).

We are also broadening the handling of BK failures, in the case
of the issue we had a ZKException.